Listview and Noitem Added.

+++ TodoApp/Views/ListView.swift
@@ -12,28 +12,37 @@ struct ListView: View {
     @EnvironmentObject var listViewModel : ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.todoItems) { item in
-                ListRowcCell(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        
+        ZStack{
+            if listViewModel.todoItems.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            }else{
+                List{
+                    ForEach(listViewModel.todoItems) { item in
+                        ListRowcCell(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
-            }
-            .onDelete(perform:listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItems)
-         }
-        .listStyle(.plain)
-        .navigationTitle("Todo List 📝")
-        .toolbar{
-            ToolbarItem(placement: .topBarLeading) {
-                EditButton()
-            }
-            ToolbarItem(placement: .topBarTrailing) {
-                NavigationLink("Add", value: "Add")
+                    .onDelete(perform:listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItems)
+                 }
+                .listStyle(.plain)
+                .toolbar{
+                    ToolbarItem(placement: .topBarLeading) {
+                        EditButton()
+                    }
+                    ToolbarItem(placement: .topBarTrailing) {
+                        NavigationLink("Add", value: "Add")
+                    }
+                }
             }
         }
+       
+        .navigationTitle("Todo List 📝")
         .navigationDestination(for: String.self) { item  in
             if item == "Add"{
                 AddView()
@@ -46,7 +55,7 @@ struct ListView: View {
     NavigationStack{
         ListView()
     }
-    //.environmentObject(ListViewModel())
+  .environmentObject(ListViewModel())
    
 }
